extract 'current_working_path' method to own module
