+++ pure/current_working_path.py
@@ -1,0 +1,7 @@
+import os
+
+from pure import colors
+
+
+def current_working_path():
+    return colors.info(os.getcwd())
